pyproject.toml
[project]
name = "edutrack-ai"
version = "0.1.0"
description = "GenAI-powered Student Submission Tracker using Django, GitHub API & Gemini AI"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "django>=4.2.7",
    "requests>=2.31.0",
    "google-generativeai>=0.3.1",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.24.0",
    "pillow>=10.1.0",
    "packaging>=25.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "flake8>=6.0.0",
    "pytest>=7.4.0",
    "pytest-django>=4.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["edutrack", "config"]

[tool.black]
line-length = 100
target-version = ['py38']
include = '\.pyi?$'

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py", "*_test.py", "tests.py"]
